rm virtual
