+++ .eslintrc
@@ -20,7 +20,8 @@
     "node_modules",
     "dist",
     "polaris-react/build",
-    "polaris-react/build-internal"
+    "polaris-react/build-internal",
+    "polaris-icons/index.ts"
   ],
   "rules": {
     "func-style": "off",
