+++ .gitignore
@@ -13,3 +13,6 @@ node_modules
 /polaris-react/build
 /polaris-react/build-internal
 /polaris-react/src/styles/_media-queries.scss
+
+# Generated file for icons build
+/polaris-icons/index.ts
