+++ .prettierignore
@@ -6,3 +6,6 @@ node_modules
 
 # temp ignore polaris.shopify.com
 /polaris.shopify.com
+
+# Ignore generated icons file
+/polaris-icons/index.ts
